Add two alternative scenarii
These scenarii promote the development of gas plants at the expense of coal plants. In the first scenario the additional capacity of the electric system is the same than in the baseline scenario and is equally distributed between coal and gas plants. In the second scenario, more relevant, the electricity generation is the same than in the baseline scenario and is equally distributed between coal and gas plants. The second scenario is added to the Makefile.